bysyscall: support libc pthreads as well as libpthread pthreads
On OL9 systems, libc delivers pthread functionality so we need
bysyscall to fall back to checking libc for instrumenting
pthread-related functions if they are not found in libpthread.so.

Also need to handle CO-RE related changes in where mm_stats are
found; in > 5.15 they are in task struct while prior to that
they are in mm_struct.  We need to ensure we can compile in both
environments, so need to define our own variants of the relevants
structs that keep the compiler happy and that CO-RE can use.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>